OSIS-7652: corrections review Bastien
