--- inscription_evaluation/templates/inscription_evaluation/blocks/soumettre_demande_inscription.html
@@ -1,31 +1,0 @@
-{% load i18n %}
-{% comment "License" %}
-    * OSIS stands for Open Student Information System. It's an application
-    * designed to manage the core business of higher education institutions,
-    * such as universities, faculties, institutes and professional schools.
-    * The core business involves the administration of students, teachers,
-    * courses, programs and so on.
-    *
-    * Copyright (C) 2015-2024 Université catholique de Louvain (http://www.uclouvain.be)
-    *
-    * This program is free software: you can redistribute it and/or modify
-    * it under the terms of the GNU General Public License as published by
-    * the Free Software Foundation, either version 3 of the License, or
-    * (at your option) any later version.
-    *
-    * This program is distributed in the hope that it will be useful,
-    * but WITHOUT ANY WARRANTY; without even the implied warranty of
-    * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    * GNU General Public License for more details.
-    *
-    * A copy of this license - GNU General Public License - is available
-    * at the root of the source code of this program.  If not,
-    * see http://www.gnu.org/licenses/.
-{% endcomment %}
-<div>
-    <a href="{% url 'inscription-evaluation:soumettre-demande-inscription' code_programme %}?sigle_formation={{ formation.sigle }}"
-       class="btn btn-primary btn-wrap btn-block-xs"
-    >
-        <i class="fa-solid fa-paper-plane"></i> {% trans 'Submit my exams enrollment' %}
-    </a>
-</div>
