+++ inscription_evaluation/templates/inscription_evaluation/ue/type_inscription.html
@@ -22,7 +22,7 @@
     * see http://www.gnu.org/licenses/.
 {% endcomment %}
 {% if session_est_egal_a_session_de_travail %}
-    <input id="{{ code_ue }}" type="checkbox" {% if inscription_evaluation %}checked{% endif %} onchange="$(this).next('span').toggle();">
+    <input id="{{ inscription.unite_enseignement.code }}" type="checkbox" {% if inscription_evaluation %}checked{% endif %} onchange="$(this).next('span').toggle();">
 {% endif %}
 {% if type_inscription == 'INSCRIPTION_PARTIELLE' %}
     <span {% if not inscription_evaluation %}style="display: none;"{% endif %} class="{% if session_est_egal_a_session_de_travail %}badge info-color{% endif %}">{{ type_inscription_txt }}</span>
